scx_utils: Add helpers for finding mount points
Move helpers from scxtop to scx_utils for finding tracefs/debugfs mount
points.
